+++ internal/sso/auth/awssso.go
@@ -22,7 +22,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
 	"strconv"
 	"sync"
 	"time"
@@ -35,6 +34,7 @@ import (
 	ssotypes "github.com/aws/aws-sdk-go-v2/service/sso/types"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
 	"github.com/davecgh/go-spew/spew"
+	"github.com/synfinatic/aws-sso-cli/internal/awsendpoint"
 	"github.com/synfinatic/aws-sso-cli/internal/awsparse"
 	ssoconfig "github.com/synfinatic/aws-sso-cli/internal/sso/config"
 	"github.com/synfinatic/aws-sso-cli/internal/sso/oidc"
@@ -411,18 +411,12 @@ func (as *AWSSSO) getRoleCredentials(accountId int64, role string, chainMap map[
 	if err != nil {
 		return storage.RoleCredentials{}, err
 	}
-	_, useFips := os.LookupEnv("AWS_USE_FIPS_ENDPOINT")
-	_, useDualStack := os.LookupEnv("AWS_USE_DUALSTACK_ENDPOINT")
 	stsSession := sts.NewFromConfig(cfg, func(o *sts.Options) {
 		if as.stsEndpoint != "" {
 			o.BaseEndpoint = aws.String(as.stsEndpoint)
 		}
-		if useFips {
-			o.EndpointOptions.UseFIPSEndpoint = aws.FIPSEndpointStateEnabled
-		}
-		if useDualStack {
-			o.EndpointOptions.UseDualStackEndpoint = aws.DualStackEndpointStateEnabled
-		}
+		o.EndpointOptions.UseFIPSEndpoint = awsendpoint.FipsEndpointState()
+		o.EndpointOptions.UseDualStackEndpoint = awsendpoint.DualStackEndpointState()
 	})
 
 	previousAccount, _ := awsparse.AccountIdToString(creds.AccountId)